Split out rating comp to shared components area
