+++ types/src/components/app-rating/index.d.ts
